HIMMC/requirements.txt
python-dotenv
flask
flask_cors
pymongo
bcrypt
google-generativeai